Companies Done

+++ Test/test station.sql
@@ -6,20 +6,15 @@
 
 declare @Tardb varchar(255) = 'DFETarget'
 declare @Sourdb varchar(255) = 'DFESource'
-declare @MainTable nvarchar(255) = 'JobOrders'
-declare @IDField nvarchar(255) = 'JobOrdersID'
-declare @DiffTableFn nvarchar(255) = 'fn_Drew_Restore_JobOrders_DiffTable_t'
-declare @RestoreTreeFn nvarchar(255) = 'fn_Drew_Restore_JobOrders_RestoreTree_t'
-declare @FindTestData bit = 0
+declare @MainTable nvarchar(255) = 'MProjects'
+declare @IDField nvarchar(255) = 'MProjectsID'
+declare @DiffTableFn nvarchar(255) = 'fn_Drew_Restore_CandidateIntroductions_DiffTable_t'
+declare @RestoreTreeFn nvarchar(255) = 'fn_Drew_Restore_CandidateIntroductions_RestoreTree_t'
+
+declare @FindTestData bit = 1
 declare @TestRestore bit = 0
+declare @MainRecordID int = 26518
 
---param
-declare @MainRecordID int = 424
-
-
-
---set @FindTestData = 1
-set @TestRestore = 1
 
 --find test data
 if @FindTestData = 1
@@ -32,7 +27,14 @@ if @TestRestore = 1
 
 --spot checks
 
---sp_Drew_Restore_Diff @SourceDBName = 'DFESource', @TargetDBName = 'DFETarget', @TableName = 'Webjobpostings', @WhereClause = N'WHERE JobOrdersID = @MainRecordID', @MainRecordID = 6407
---SELECT * FROM Drew_Diff
---DROP TABLE Drew_Diff
+declare @tablename varchar(255) = null
+
+--set @tablename = 'ProjectsCandidateBlocks'
+
+if @tablename is not null begin
+	declare @WhereClause nvarchar(max) = (select top 1 WhereClause from dbo.fn_Drew_Restore_CandidateIntroductions_DiffTable_t() where TableName = @tablename)
+	exec sp_Drew_Restore_Diff @SourceDBName = 'DFESource', @TargetDBName = 'DFETarget', @TableName = @tablename, @WhereClause = @WhereClause, @MainRecordID = 26518
+	SELECT * FROM Drew_Diff
+	DROP TABLE Drew_Diff
+end
